\encoding{latin1}
\name{plotfp}
\alias{plotfp}
\title{Factorial Planes from Coordinates}
\description{
 It plots factorial planes from a coordinate table
}
\usage{
plotfp(co,x=1,y=2,eig=NULL,cal=NULL,ucal=0,xlim=NULL,ylim=NULL,main=NULL,
                   rotx=FALSE,roty=FALSE,eti=row.names(co),
                        axislabel=TRUE,col.row="black",cex=0.8,cex.row=0.8,
                        all.point=TRUE,cframe=1.2,cex.global=1,infaxes="out")
}
\arguments{
  \item{co}{matrix or data.frame with coordinates}
  \item{x}{the component like horizontal axis}
  \item{y}{the component like vertical axis}
  \item{eig}{numeric with the eigenvalues}
  \item{cal}{matrix or data.frame with the square cosinus}
  \item{ucal}{quality representation threshold (percentage) in the plane . Default ucal=0}
  \item{xlim}{the x limits (x1, x2) of the plot}
  \item{ylim}{the y limits of the plot}  
  \item{main}{graphic title}
  \item{rotx}{TRUE if you want change the sign of the horizontal coordinates. Default FALSE }
  \item{roty}{TRUE if you want change the sign of the vertical coordinates. Default FALSE }
  \item{eti}{selected row points for the graphic. Default all points}
  \item{axislabel}{if it is TRUE the axis information is written}
  \item{col.row}{color for row points and row labels. Default "black"}
  \item{cex}{global scale for the labels. Default cex=0.8}
  \item{cex.row}{scale for row points and row labels. Default cex.row=0.8 }
  \item{all.point}{If if is TRUE, all points are outlined. Default all.point=TRUE }
  \item{cframe}{scale for graphic limits}
  \item{cex.global}{scale for the label sizes }
  \item{infaxes}{place to put the axes information: "out","in","no". Default infaxes="out".
    If infaxes="out" the graphic is similar to \code{FactoMineR} graphics, otherwise the style
    is similar to the one in \code{ade4}, without axes information when infaxes="no"}
}
\details{
  Plot the selected factorial plane. 
}
\value{
  It graphs the factorial plane x,y using  co and optional information of eigenvalues and representation quality of the points.
  If ucal > 0, only the points with  the quality of representation 
  on the plane bigger than ucal are pointed
}
\author{Campo Elias Pardo \email{cepardot@unal.edu.co} 

  \url{http://www.docentes.unal.edu.co/cepardot}
}
\examples{
data(mores5)
attach(mores5)
ca <- dudi.coa(tab,scannf=FALSE,nf=4)
wibca <- witwit.model(ca,rbl5,cbl5,model="B",weight="mfa",scannf=FALSE,nf=4)
detach(mores5)
mfa <- wibca2mfa(wibca)
attach(mfa)
plotfp(colb$coor,colb$eig,main="WIBCA Column bands like MFA of Spain Mortality, First Factorial Plane")
dev.new()
plotfp(rowb$coor,rowb$eig,main="WIBCA Row bands like MFA of Spain Mortality, First Factorial Plane")
}
\keyword{multivariate}
\keyword{hplot}

